move non-selectors out to utilty; adjust reload timer
