+++ common/components/WalletDecrypt/components/Trezor.tsx
@@ -5,11 +5,12 @@ import TrezorConnect from 'vendor/trezor-connect';
 import DeterministicWalletsModal from './DeterministicWalletsModal';
 import './Trezor.scss';
 import { Spinner } from 'components/ui';
-import { getSingleDPath, getNetworkConfig, getPaths } from 'selectors/config';
+import { getNetworkConfig } from 'selectors/config';
 import { AppState } from 'reducers';
 import { connect } from 'react-redux';
 import { SecureWalletName } from 'config';
 import { DPath } from 'config/dpaths';
+import { getPaths, getSingleDPath } from 'utils/network';
 
 //todo: conflicts with comment in walletDecrypt -> onUnlock method
 interface OwnProps {
